% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipseConstructor.R
\name{ellipseConstructor}
\alias{ellipseConstructor}
\title{Extracts x- and y- data for visualization of photoreceptor dependencies from a sensitivity list}
\usage{
ellipseConstructor(path = ".")
}
\description{
Extracts x- and y- data for visualization of photoreceptor dependencies from a sensitivity list
}
\examples{
\dontrun{
xyData <- ellipseConstructor(examplePath)
eModel <- fitModel(xyData)
print(eModel$results$alphaEstimate)
plotEllipse(xyData, ellipseModel = eModel)
}
}
