% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLEDContrast.R
\name{findLEDContrasts}
\alias{findLEDContrasts}
\title{Calculate LED contrasts}
\usage{
findLEDContrasts(pursuedContr, lmean, ConeFund, maxContrast = T)
}
\arguments{
\item{lmean}{time-averaged luminances of the LEDs}

\item{ConeFund}{cone funamentals to be used}

\item{maxContrast}{use the maximal contrasts at the photoreceptor levels that can be achieved technically while keeping the contrast ratios from pursuedContrast constant?}

\item{pursuedContrast}{the contrast in the photoreceptors that should be achieved}
}
\value{
LED contrasts that result in the desired contrasts at the photoreceptor level
}
\description{
Calculates the LED contrast that result in a pursued contrast at the photoreceptor level, given certain time-averaged LED illuminances and given cone fundamentals.
}
\examples{
\dontrun{
#### set given values
## mean luminance of primaries
lmean <- c(red = 20, green = 40, blue = 3, cyan = 20)
## create M-cone isolating stimlui
pursuedContr <- c(rod = 0, scone = 0, mcone = 20, lcone = 0)
## Find LED contrast that result in pursued contrasts
LEDcontrast100 <- findLEDContrasts(pursuedContr, lmean, ConeFund)
# Calculate the maximal photoreceptor contrast this results in
maxContrast <- findPhotoreceptorContrasts(LEDcontrast100, lmean, ConeFund)
createPresetFile("mcone", "außen", LEDcontrast100, c(20, 40, 3, 20), lmean)
}
}
