% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPhotoreceptorContrasts.R
\name{findPhotoreceptorContrasts}
\alias{findPhotoreceptorContrasts}
\title{Calculates the contrasts at the photoreceptor levels for given contrasts and mean luminances of the LED}
\usage{
findPhotoreceptorContrasts(LEDcontrast, lmean, ConeFund)
}
\arguments{
\item{LEDcontrast}{a vector of LED contrasts}

\item{lmean}{mean luminances of the LEDS}

\item{ConeFund}{cone fundamentals to be used}
}
\value{
a vector with the contrasts at the photoreceptor level
}
\description{
For a given set of LED mean luminances and LED contrast, contrasts at the photoreceptor level can be calculated when the cone fundamentals are known.
}
