% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEllipse.R
\name{fitEllipse}
\alias{fitEllipse}
\title{Fits an ellipse model to data from getPhotoreceptorCoordinates}
\usage{
fitEllipse(photoreceptorCoordinates, x = "lcone", y = "mcone", a1 = 1,
  a2 = 1, alpha = 1)
}
\description{
Fits an ellipse model to data from getPhotoreceptorCoordinates
}
\examples{
\dontrun{
getPhotoreceptorCoordinates(path, ConeFundamentals10) \%>\%
 filter(patid == 1, seye == "OD", frequency == 2) \%>\%
 fitEllipse() \%>\%
 plotEllipse()
}
}
