% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoreceptorCoordinates.R
\name{getPhotoreceptorCoordinates}
\alias{getPhotoreceptorCoordinates}
\title{Calculates the contrasts at threshold for the different photoreceptor levels}
\usage{
getPhotoreceptorCoordinates(path = ".", ConeFund = NA)
}
\arguments{
\item{path}{path to a directory containing multiple result files}

\item{ConeFund}{cone fundamentals to be used}
}
\value{
a table with contrasts at threshold at the photoreceptor level
}
\description{
The contrasts for all four photoreceptors at threshold are calculated from severel result files. This function does not rely on known presets.
}
