% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResultFileList.R
\name{getResultFileList}
\alias{getResultFileList}
\title{Get a list of result files in the directory "path"}
\usage{
getResultFileList(path = ".")
}
\arguments{
\item{path}{Directory that is to be searched for result files}
}
\value{
a list of file names
}
\description{
Get a list of result files in the directory "path"
}
\author{
Cord Huchzermeyer
}
