% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoreceptorCoordinates.R
\name{read.flimmerkiste}
\alias{read.flimmerkiste}
\title{Alternative method for reading result files.}
\usage{
read.flimmerkiste(filename, ConeFund)
}
\arguments{
\item{filename}{path to a result files}

\item{ConeFund}{cone fundamentals to be used}
}
\value{
a table with contrasts at threshold at the photoreceptor level
}
\description{
THis function reads a results file and calculates photoreceptor contrasts.
}
