% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultFileL4S.R
\name{resultFile}
\alias{resultFile}
\title{Read a single result file}
\usage{
resultFile(name, presets)
}
\arguments{
\item{name}{name of the file}

\item{presets}{presets that are used for identifying conditions and calculating sensitivities}
}
\value{
an object of the type "resultFile"
}
\description{
Read a single result file
}
