% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultListL4S.R
\name{resultList}
\alias{resultList}
\title{Loads examinations files from a path and creates an object resultList}
\usage{
resultList(pfad = ".")
}
\arguments{
\item{path}{path that contains multiple resultFile files}
}
\value{
an object of the type "resultList"
}
\description{
Loads examinations files from a path and creates an object resultList
}
