% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityList.R
\name{sensitivityList.character}
\alias{sensitivityList.character}
\title{Creates a sensitivity list from raw data}
\usage{
\method{sensitivityList}{character}(path = ".")
}
\arguments{
\item{path}{directory containing a number of txt-files with temporal contrast sensitivity measurements}
}
\description{
Creates a sensitivity list from raw data
}
