% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityList.R
\name{sensitivityList.default}
\alias{sensitivityList.default}
\title{Creates a sensitivity list from raw data}
\usage{
\method{sensitivityList}{default}(x)
}
\arguments{
\item{x}{any object}
}
\description{
Creates a sensitivity list from raw data
}
\details{
objects with a class other than character or resultList result only in a message that there is no method "sensitivityList" for this class.
}
