% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityList.R
\name{sensitivityList}
\alias{sensitivityList}
\title{Creates a sensitivity list from raw data}
\usage{
sensitivityList(x)
}
\arguments{
\item{x}{either a string containing a path or a resultList-object}
}
\description{
Creates a sensitivity list from raw data
}
\seealso{
sensitivityList.character, sensitivityList.resultList
}
