% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityList.R
\name{sensitivityList.resultList}
\alias{sensitivityList.resultList}
\title{Creates a sensitivity list from raw data}
\usage{
\method{sensitivityList}{resultList}(resultList)
}
\arguments{
\item{resultList}{an object of class resultList}
}
\description{
Creates a sensitivity list from raw data
}
\examples{
\dontrun{
rl <- resultList(".")
sl <- sensitivityList(rl)
print(sl)
}
}
