% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrastFunctions.R
\name{findLEDContrasts}
\alias{findLEDContrasts}
\title{Calculate LED contrasts}
\usage{
findLEDContrasts(pursuedContr, lmean, ConeFund, maxContrast = T)
}
\description{
Calculate LED contrasts
}
\examples{
# use conversion factors from spreadsheet
lum2nrgB <- c(red = 0.0216,
             green = 0.00155,
             blue = 0.0242,
             cyan = 0.0024)
## how to convert from luminance to energy: select method A or B
lum2nrg <- lum2nrgB
#### set given values
## mean luminance of primaries
lmean <- c(red = 20,
          green = 40,
          blue = 3,
          cyan = 20)
## pursued receptor contrast, actual value irrelevant
pursuedContr <- c(rod = 0,
                 scone = 0,
                 mcone = 20,
                 lcone = 0)
# Find LED contrast that result in pursued contrasts
LEDcontrast100 <- findLEDContrasts(pursuedContr, lmean, ConeFund)
# Calculate the maximal photoreceptor contrast this results in
maxContrast <- findPhotoreceptorContrasts(LEDcontrast100, lmean, ConeFund)
print(round(maxContrast, 2))
Kliste <-c()
pursuedContr <- c(rod = 0,
                scone = 0,
                 mcone = 10,
                 lcone = 0)
LEDcontrast100 <- findLEDContrasts(pursuedContr,
                                  lmean,
                                  ConeFund,
                                  T)
createPresetFile("e_L0Mp1", "außen", LEDcontrast100, c(20, 40, 3, 20), lmean)
maxContrast <- findPhotoreceptorContrasts(LEDcontrast100, lmean, ConeFund)
Kliste <- c(Kliste, max(maxContrast))
}
