% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultFileL4S.R
\name{resultFile}
\alias{resultFile}
\title{Read a single result file}
\usage{
resultFile(name, presets = flickerbox::presets)
}
\arguments{
\item{name}{name of the file}

\item{presets}{presets that are used for identifying conditions and calculating sensitivities}
}
\value{
an object of the type "resultFile"
}
\description{
Use this function to read a single results file. Such a file contains all stimuli presented and two thresholds (one for the up-, the other for the down-staircase).
}
\examples{
# Get the maximal possible contrast at the photoreceptor level for this measurement (instrument gamut)
maxContrast <- resultFile("pathToFile")$maxContrast
}
